+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/tests/latest/test_network_commands.py
@@ -70,8 +70,7 @@ class NetworkLoadBalancerWithSku(ScenarioTest):
         ])
 
 
-# TODO: revert to ScenarioTest once #9401 is addressed.
-class NetworkPrivateEndpoints(LiveScenarioTest):
+class NetworkPrivateEndpoints(ScenarioTest):
 
     @ResourceGroupPreparer(name_prefix='cli_test_network_private_endpoints')
     def test_network_private_endpoints(self, resource_group):